fix: move components to a different folder
