+++ src/components/index.ts
@@ -1,7 +1,5 @@
 import { default as Header } from './Header/Header';
 import { default as Footer } from './Footer/Footer';
-import { default as AnswerModal } from './AnswerModal/AnswerModal';
-import { default as Results } from './Results/Results';
 import { default as Button } from './Button/Button';
 
-export { Header, Footer, AnswerModal, Results, Button };
+export { Header, Footer, Button };
