+++ src/views/GamePage/components/AnswerModal/AnswerModal.css
