+++ src/views/GamePage/components/AnswerModal/AnswerModal.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Button } from '..';
+import { Button } from '../../../../components';
 
 import './AnswerModal.css';
