+++ src/views/GamePage/components/AnswerModal/index.ts
@@ -1,0 +1,3 @@
+import { default as AnswerModal } from './AnswerModal';
+
+export { AnswerModal };
