+++ src/views/GamePage/components/GameEngine/index.ts
@@ -1,0 +1,3 @@
+import { default as GameEngine } from './GameEngine';
+
+export { GameEngine };
