+++ src/views/GamePage/components/NavigationBtns/index.ts
@@ -1,0 +1,3 @@
+import { default as NavigationBtns } from './NavigationBtns';
+
+export { NavigationBtns };
