+++ src/views/GamePage/components/Results/Results.css
