+++ src/views/GamePage/components/Results/components/Result/Result.css
