+++ src/views/GamePage/components/Results/components/Result/Result.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Picture } from '../../../../types';
+import { Picture } from '../../../../../../types';
 
 import './Result.css';
