+++ src/views/GamePage/components/Results/components/index.ts
