+++ src/views/GamePage/components/Results/index.ts
@@ -1,0 +1,3 @@
+import { default as Results } from './Results';
+
+export { Results };
