+++ src/views/GamePage/components/RoundInfo/RoundInfo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Button } from '..';
 
+import { Button } from '../../../../components';
 import './RoundInfo.css';
 
 const RoundInfo = ({
